requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
pre-commit>=3.4.0

# Documentation
Sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0